+++ tests/test_base.py
@@ -1,52 +1,16 @@
 import pytest
-from bgameb.base import Components, CardTexts
-from bgameb.stuff import RollerType, CardType, Roller, Card, BaseStuff
+from bgameb.base import Components
+from bgameb.stuff import Roller, Card, BaseStuff
 from bgameb.errors import ComponentNameError
-
-
-class TestCardText:
-    """Test CardText class
-    """
-
-    def test_card_text_operations(self) -> None:
-        """Test get, set and delete operations of
-        CardText
-        """
-        text = CardTexts()
-        text.this = 'this'
-        assert text.this == 'this', 'not set or cant get'
-        assert text.__repr__() == "CardTexts(this='this')", 'wrong repr'
-        text.this = 'that'
-        assert text.this == 'that', 'not set or cant update'
-        del text.this
-        with pytest.raises(
-            AttributeError, match='this'
-            ):
-            text.this
-        with pytest.raises(
-            KeyError, match='this'
-            ):
-            del text.this
-
-    def test_card_text_equal(self) -> None:
-        """Test equal of CardTexts
-        """
-        text1 = CardTexts()
-        text2 = CardTexts()
-        assert text1 == text2, 'not equal'
-        text2.this = 'this'
-        assert text1 != text2, 'equal'
 
 
 class TestComponents:
     """Test CardText class
     """
     components = [
-        (RollerType, 'dice'),
-        (CardType, 'card'),
         (Roller, 'dice'),
         (Card, 'card'),
-    ]
+        ]
 
     @pytest.mark.parametrize("_class, name", components)
     def test_components_access_to_attr(self, _class, name: str) -> None:
@@ -127,7 +91,7 @@ class TestComponents:
             match='is exist in'
         ):
             comp._add(_class, name=name)
-        comp._add(RollerType, name='this_is')
+        comp._add(Roller, name='this_is')
         assert comp.this_is, 'component not added'
         with pytest.raises(
             ComponentNameError,